github_app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
PyGithub==2.1.1
PyJWT==2.8.0
python-dotenv==1.0.1 